Completed section 15
Covered the differences between ES5 functions and arrow functions inside
objects. A ES5 method defined in an object property will have 'this' bound
to the object. Arrow functions do not bind 'this' instead the inherit
'this' from the context they were created. Anonymous ES5 functions are
not bound to an object if they exist inside a method.

Reviewed ES6 method definition syntax.